Create functions for get source files
Create new function for getting path
of file and function for copy file to
actual directory. For getting path of file is used
limeGetFilePath() function.
For copy of file from default location
to specific dir is used limeSourceFile().
Also use this new functions for tests,
where is it usefull.

Fix issue #119