Basic attestation scenario - agent,verifier, registrar in container
Basic attestation scenario, where agent,verifier and registrar
are in containers. Small changes in limeLib in name of function
for working with containers and removing duplicated option from
function. Also create function to prepare and run registrar and verifier
containers. Change of location where is set permission for access to
tpm to lime helper function and updating in older container tests.