Add simple sanity check for durable attestation
This test was pieced together with examples from
measured-boot-swtpm-sanity and basic-attestation-with-ima-signatures.
The main idea here is to add an agent with both measured boot and
runtime policies, and make sure offline attestation works on it.

When `keylime_attest` is run, all that matters is the exit code (0 will
indicate that all stored attestation artifacts were properly attested)

Signed-off-by: Marcio Silva <marcio.a.silva@ibm.com>