Library: Use entrypoint to chown certificate dir for agent
Use --entrypoint to override the image entrypoint in case it has an
entrypoint set.

The agent needs the certificate directory to be accessible by the
internal 'keylime' user in order to drop privileges inside the
container.  For this, it is necessary for the files owner uid to match
the internal 'keylime' uid.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>