Provide public key for signature verification on command line
Keylime does not allow the trivial passing of signature verification on
the self-signed certificate embedded in the DSSE envelope of a signed
policy anymore. Instead, it requires that a public key for signature
verification be given by the user. Therefore, pass the public key needed
for signature verification via the command line.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>