Use the latest old supported version that is not the current version
This change is to cover a corner case when the python components API
version is bumped before the agent API version is bumped.

In such a case, the latest old supported API version for the verifier
would be the current agent API version. Then it is necessary to use the
second to last to cause a difference in the API version.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>