Add test case for system installed CA certificate for webhook
Add a test case where the CA certificate of the revocation notification
webhook certificate issuer is installed in the system-wide trust store
and not provided in the 'trusted_server_ca' configuration option.

This is to check that the system-wide trust store is considered when
verifying the revocation notification webhook certificate.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>