library: improve how lime library checks for keylime-related processes
At some point we need to check for keylime processes running so that
we can e.g. stop them, but the current approach is doing a too broad
search that it kills unrelated processes as well

As an example, if we have journalctl watching e.g. the verifier logs
and we try to stop the verifier, the journalctl process would also be
killed, which is not what we want.

The updated code tries to be more fine-grained and appears to do the
right thing(tm).

Signed-off-by: Sergio Correia <scorreia@redhat.com>