Add functional/add-agent-with-malformed-ek-cert
In this test, swtpm will provide a properly signed (by the swtpm-local) EK
certificate that happens to be considered malformed by python-cryptoghraphy,
due to its stricter ASN.1 parsing since version 35; OpenSSL considers this
certificate valid, as it is more lenient with its ASN.1 parsing.

The test will check that we are still able to add such and agent that has an
EK certificate with these characteristics.

Signed-off-by: Sergio Correia <scorreia@redhat.com>