Check that cached malformed EK cert does not affect other certs
In the fix attempt https://github.com/keylime/keylime/pull/1794, the
cache was incorrectly added to the model object instead of the
certificate object instance, meaning that the cached certificate would
be used for all following certificates.

This adds a check to verify that the cached malformed certificate will
not affect following certificates.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>