Use custom certificates for the agent too
Previously, the agent supported only RSA key/certificate as the mTLS
key because it was also used to encrypt data in the secure payload
mechanism.

With https://github.com/keylime/rust-keylime/pull/1129, the agent is now
able to use different algorithms for mTLS.

When testing PQC algorithms, set ECDSA keys for the agent due to
https://issues.redhat.com/browse/RHEL-117439. Once this is fixed, the
special treatment for the agent mTLS should be removed.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>