Add test vyrifying that registrar rejects duplicate UUID [WIP]
Test adds an agent with UUID and attemps to add another agent with identical UUID and chcekc if registrar explicitely refuses it.
It is prepared for both pull and push mode attestation.

WIP: Currently the behaviour seems not to be as expected, registrar accepts agent with duplicite UUID - test needs review.